Graph coloring register allocation (#42)
* change parent to dominator

* impl simple single-bb igraph construction

* interference graph sanity check test

* inter-block ig construction except loops

* more ig construction sanity checks

* impl ig construction for loops

* basic clustering algorithm

* remove old src directory

* get ig construction up to speed w/ new ir stuff

* infra setup for final coloring allocator impl

* finished basic graph coloring for regalloc

* update regalloc tests for costs & entries

* finished graph coloring register alloc

+++ tinyc/src/codegen/dlx/utils.rs
@@ -13,7 +13,7 @@ macro_rules! do_with_store {
 
         // emit the store
         if let Some(offset) = offset {
-            $self.emit_store($reg, Register::RFP, $self.stack_offset(offset as i16));
+            $self.emit_store($reg, Register::RFP, $self.stack_offset(offset));
         }
     }
 }
